refactor: split login, signup, logout util folder
