+++ tests/Feature/Auth/LoginTest.php
@@ -3,7 +3,7 @@
 namespace Tests\Feature\Auth;
 
 use Tests\TestCase;
-use Tests\Util\Auth\LoginUtil;
+use Tests\Util\Auth\Login\LoginUtil;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 
 class LoginTest extends TestCase
