+++ tests/Util/Auth/Login/InvalidLoginCredentials.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Util\Auth;
+namespace Tests\Util\Auth\Login;
 
 class InvalidLoginCredentials
 {
