+++ tests/Util/Auth/Signup/SignupUtil.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Util\Auth;
+namespace Tests\Util\Auth\Signup;
 
 class SignupUtil
 {
